validate: Split clusters flow tests to clusters_test.go
Move all TestValidateClusters* tests and clusters-specific mock
instances (clustersGatherDataFailed, checkS3Failed, checkS3Canceled)
from command_test.go to clusters_test.go. command_test.go now contains
only the common TestValidatedDeleted test (75 lines).
